+++ pkg/config/deprecated.go
@@ -20,7 +20,7 @@ var deprecatedOpts = map[string]deprecated{
 	"dumpdir":       unsupported,
 	"debug-updated": {"Use '-d' as option instead.", nil},
 	"execurl":       unsupported,
-	"filter":        unsupported,
+	"filter":        {"Use 'item-filter' instead.", nil},
 	"disable-ssl-verification": {"Interpreted as 'tls-no-verify'.", func(i interface{}, global *GlobalOptions, opts *Options) {
 		val, ok := i.(bool)
 		if ok {